refactor: move to new actions folder
